Update non teaching staff users.py

+++ non teaching staff users.py
@@ -1,4 +1,4 @@
-a=float(input("Enter Total Number of Users : "))
-b=float(input("Enter Number of Staff Users : "))
+a=int(input("Enter Total Number of Users : "))
+b=int(input("Enter Number of Staff Users : "))
 c=a-b-(b//3)
 print("Student Users Are : " , c)
